separação dos arquivos de jogo
