+++ lib/game/flappy_game.dart
@@ -1,0 +1,25 @@
+import 'package:flame/game.dart';
+import 'package:flutter/material.dart';
+
+class FlappyGame extends FlameGame {
+  @override
+  Future<void> onLoad() async {
+    super.onLoad();
+    debugPrint('Jogo carregado');
+  }
+
+  @override
+  void update(double dt) {
+    super.update(dt);
+    debugPrint('Update: $dt');
+  }
+
+  @override
+  void render(Canvas canvas) {
+    super.render(canvas);
+    Paint cor = Paint();
+    cor.color = Colors.white;
+
+    canvas.drawCircle(Offset(size.x / 2, size.y / 2), 100, cor);
+  }
+}
